pyproject.toml
[project]
name = "01-board-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "beautifulsoup4==4.14.3",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6",
    "distro==1.9.0",
    "et-xmlfile==2.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jiter==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.0",
    "langchain-core==1.2.4",
    "langchain-ollama>=1.0.1",
    "langchain-openai==1.1.6",
    "langgraph==1.0.5",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.3.1",
    "langsmith==0.5.0",
    "lxml==6.0.2",
    "numpy==2.4.0",
    "openai==2.14.0",
    "opendartreader==0.2.3",
    "openpyxl==3.1.5",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "packaging==25.0",
    "pandas==2.3.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-file==3.0.1",
    "requests-toolbelt==1.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8.1",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.2",
    "uuid-utils==0.12.0",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
    "psycopg2-binary>=2.9.10",
]
